requirements.txt
streamlit
ultralytics
pillow
numpy<2